pyproject.toml
[tool.poetry]
name = "test-for-colab"
version = "0.1.0"
description = ""
authors = ["KentaroKajiyama <kemushi555@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
numpy = "^2.2.2"
rich = "^13.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "test_for_colab.main:main"